WIP nw

+++ src/main/y/sort/network/sort11.hpp
@@ -5,13 +5,13 @@
 namespace upsylon {
 		//! network sort for 11 items
 	template <> struct network_sort<11> {
-		static const size_t I[35]; //!< I swap indices
-		static const size_t J[35]; //!< J swap indices
+		static const unsigned I[35]; //!< I swap indices
+		static const unsigned J[35]; //!< J swap indices
 		//! sort...
 		template <typename T,typename COMPARE> static inline
 		void on(T *a, COMPARE &compare) throw() {
 			assert(NULL!=a);
-			for(size_t k=0;k<35;++k){
+			for(unsigned k=0;k<35;++k){
 				T &aI=a[I[k]], &aJ=a[J[k]];
 				if(compare(aJ,aI)<0) bswap<T>(aI,aJ);
 			}
@@ -22,8 +22,8 @@ namespace upsylon {
 			assert(NULL!=a);
 			assert(NULL!=b);
 			for(size_t k=0;k<35;++k){
-				const size_t  i = I[k],   j = J[k];
-				T           &aI = a[i], &aJ = a[j];
+				const unsigned  i = I[k],   j = J[k];
+				T             &aI = a[i], &aJ = a[j];
 				if(compare(aJ,aI)<0) { bswap<T>(aI,aJ); bswap<U>(b[i],b[j]); }
 			}
 		}
