+++ src/main/y/sort/network/sort2.cpp
@@ -1,9 +1,9 @@
 #include "sort2.hpp"
 namespace upsylon {
-	const size_t network_sort<2>::I[1] = {
+	const unsigned network_sort<2>::I[1] = {
  0};
 
-	const size_t network_sort<2>::J[1] = {
+	const unsigned network_sort<2>::J[1] = {
  1};
 
 }
