+++ src/main/y/sort/network/sort5.cpp
@@ -1,9 +1,9 @@
 #include "sort5.hpp"
 namespace upsylon {
-	const size_t network_sort<5>::I[9] = {
+	const unsigned network_sort<5>::I[9] = {
  0, 3, 2, 2, 1, 0, 0, 1, 1};
 
-	const size_t network_sort<5>::J[9] = {
+	const unsigned network_sort<5>::J[9] = {
  1, 4, 4, 3, 4, 3, 2, 3, 2};
 
 }
