+++ src/main/y/sort/network/sort6.cpp
@@ -1,9 +1,9 @@
 #include "sort6.hpp"
 namespace upsylon {
-	const size_t network_sort<6>::I[12] = {
+	const unsigned network_sort<6>::I[12] = {
  1, 4, 0, 3, 0, 3, 2, 0, 1, 2, 1, 2};
 
-	const size_t network_sort<6>::J[12] = {
+	const unsigned network_sort<6>::J[12] = {
  2, 5, 2, 5, 1, 4, 5, 3, 4, 4, 3, 3};
 
 }
