+++ src/main/y/sort/network/sort7.cpp
@@ -1,9 +1,9 @@
 #include "sort7.hpp"
 namespace upsylon {
-	const size_t network_sort<7>::I[16] = {
+	const unsigned network_sort<7>::I[16] = {
  1, 3, 5, 0, 3, 4, 0, 4, 2, 0, 1, 0, 2, 1, 2, 2};
 
-	const size_t network_sort<7>::J[16] = {
+	const unsigned network_sort<7>::J[16] = {
  2, 4, 6, 2, 5, 6, 1, 5, 6, 4, 5, 3, 5, 3, 4, 3};
 
 }
